chatterpy_requirements.txt
langchain-community==0.4.1
torch==2.9.1
ibm_watsonx_ai==1.4.11
langchain-ibm==1.0.2
pypdf==6.5.0
Wikipedia-API==0.8.1
langchain-ollama==1.0.1
qdrant-client==1.16.2
langchain-qdrant==1.1.0
streamlit==1.52.2
python-dotenv==1.2.1
langchain-openai==1.1.6